+++ src/App.js
@@ -1,6 +1,6 @@
 import Home from "./pages/home";
 import Login from "./pages/login";
-import Profile from "./pages/profile/profile";
+import Profile from "./pages/profile";
 import Register from "./pages/register";
 import {
   BrowserRouter as Router,
