+++ src/components/feed.jsx
@@ -9,11 +9,14 @@ const Feed = ({ username, profile }) => {
   const [posts, setPosts] = useState([]);
   const { user } = useContext(AuthContext);
 
+  // api endpoint
+  const baseUrl = process.env.REACT_APP_BASE_URL;
+
   useEffect(() => {
     const fetchPosts = async () => {
       const res = username
-        ? await axios.get("/posts/profile/" + username)
-        : await axios.get(`/posts/timeline/${user._id}`);
+        ? await axios.get(baseUrl + "/api/posts/profile/" + username)
+        : await axios.get(`${baseUrl}/api/posts/timeline/${user._id}`);
       setPosts(
         res.data.sort((p1, p2) => {
           return new Date(p2.createdAt) - new Date(p1.createdAt);
@@ -22,7 +25,7 @@ const Feed = ({ username, profile }) => {
     };
 
     fetchPosts();
-  }, [user._id, username]);
+  }, [user._id, username, baseUrl]);
 
   return (
     <div className="md:flex-[6] flex-1">
