--- src/pages/profile/profile.css
@@ -1,52 +1,0 @@
-.profileContainer{
-    display: flex;
-}
-
-.profileRight{
-    flex: 9;
-}
-
-
-.profileCover{
-    height: 320px;
-    position: relative;
-}
-
-.profileCoverImage{
-    width: 100%;
-    height: 250px;
-    object-fit: cover;
-}
-
-.profileUserImage{
-    width: 150px;
-    height: 150px;
-    border-radius: 50%;
-    object-fit: cover;
-    position: absolute;
-    left: 0;
-    right: 0;
-    margin: auto;
-    top: 170px;
-    border: 3px solid white;
-}
-
-.profileInfo{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin-top: 10px;
-}
-
-.profileInfoName{
-    font-size: 24px;
-}
-
-.profileInfoDesc{
-    font-weight: 300;
-}
-
-.profileRightBottom{
-    display: flex;
-}
