--- src/pages/profile/profile.jsx
@@ -1,62 +1,0 @@
-import axios from "axios";
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import Feed from "../../components/feed";
-import Leftbar from "../../components/leftbar/leftbar";
-import Rightbar from "../../components/rightbar/rightbar";
-import TopBar from "../../components/topbar";
-import "./profile.css";
-
-const Profile = () => {
-  const PF = process.env.REACT_APP_PUBLIC_FOLDER;
-  const [user, setUser] = useState({});
-  const { username } = useParams();
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const res = await axios.get(`/users/?username=${username}`);
-      setUser(res.data);
-    };
-
-    fetchUser();
-  }, [username]);
-
-  return (
-    <>
-      <TopBar />
-      <div className="profileContainer">
-        <Leftbar />
-        <div className="profileRight">
-          <div className="profileRightTop">
-            <div className="profileCover">
-              <img
-                className="profileCoverImage"
-                src={user.coverPicture || PF + "person/noCover.jpeg"}
-                alt=""
-              />
-              <img
-                className="profileUserImage"
-                src={
-                  user.profilePicture
-                    ? PF + user.profilePicture
-                    : PF + "/person/10.png"
-                }
-                alt=""
-              />
-            </div>
-            <div className="profileInfo">
-              <h4 className="profileInfoName">{user.username}</h4>
-              <span className="profileInfoDesc">{user.desc}</span>
-            </div>
-          </div>
-          <div className="profileRightBottom">
-            <Feed username={username} />
-            <Rightbar profile user={user} />
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default Profile;
